database/requirements.txt
dotenv==0.9.9
flask==3.1.1
fpdf2==2.8.2
openpyxl==3.1.5
pandas==2.2.3
psycopg2==2.9.10
python-docx==1.1.2
supabase==2.15.2